implemented art for dialogue scene

+++ Assets/Scripts/Core/PersistentGameManager.cs
@@ -222,7 +222,7 @@ namespace Core
             _loading = false;
         }
 
-        public float SusMeter { get ; set ; }
+        public int SusMeter { get ; set ; }
 
         public PassportData PassportData { get; set; }
         public BombData BombData { get; set; }
